+++ src/templating.ts
@@ -1,12 +1,20 @@
+const FILE_COUNT_EXP = /\{\{\s*fileCount\s*\}\}/;
 const FILES_LIST_EXP = /\{\{\s*filesList\s*\}\}/;
 const OPTIMIZED_COUNT_EXP = /\{\{\s*optimizedCount\s*\}\}/;
+const SVG_COUNT_EXP = /\{\{\s*svgCount\s*\}\}/;
 
 const format = {
+  fileCount: (template: string, value: number): string => {
+    return template.replace(FILE_COUNT_EXP, value.toString());
+  },
   filePaths: (template: string, value: string[]): string => {
     return template.replace(FILES_LIST_EXP, "- " + value.join("\n- "));
   },
   optimizedCount: (template: string, value: number): string => {
     return template.replace(OPTIMIZED_COUNT_EXP, value.toString());
+  },
+  svgCount: (template: string, value: number): string => {
+    return template.replace(SVG_COUNT_EXP, value.toString());
   },
 };
 
@@ -26,8 +34,10 @@ function formatAll(
 
 
 export type CommitData = {
+  fileCount: number;
   filePaths: (string | undefined)[];
   optimizedCount: number;
+  svgCount: number;
 }